Safari-609.1.11.1/LayoutTests/http/tests/websocket/tests/hybi/handshake-fail-by-mismatch-protocol-header-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-mismatch-protocol-header' failed: Error during WebSocket handshake: Sec-WebSocket-Protocol mismatch
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-mismatch-protocol-header' failed: Error during WebSocket handshake: Sec-WebSocket-Protocol mismatch
CONSOLE MESSAGE: WebSocket connection to 'ws://localhost:8880/websocket/tests/hybi/handshake-fail-by-mismatch-protocol-header' failed: Error during WebSocket handshake: Sec-WebSocket-Protocol mismatch
Test whether WebSocket handshake fails if the server sends mismatched Sec-WebSocket-Protocol header.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS closeEvent.wasClean is false
PASS closeEvent.wasClean is false
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
